Speed up type checking
Eliminate unnecessary function bodies to speed up the type checker.

The tests run slower because the act of loading the package modifies the AST in a way specific to the query, but in real-world usage this is 2-3x faster.

+++ main_test.go
@@ -55,7 +55,7 @@ func TestRunOutsideGopath(t *testing.T) {
 	}
 
 	tests := []struct {
-		Pos int64
+		Pos int
 		Doc string
 	}{
 		{Pos: 23, Doc: "\tPackage fmt implements formatted I/O"},    // import "fmt"
